retarded shit requirements by prof

+++ core/src/com/project/game/Entity/CharacterEntity.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.math.Rectangle;
 
 public abstract class CharacterEntity extends Entity {
     // to do add health and what not
-    public CharacterEntity(int entityID, int posX, int posY, String type, Texture tex, Rectangle rec) {
+    public CharacterEntity(int entityID, float posX, float posY, String type, Texture tex, Rectangle rec) {
         super(entityID, posX, posY, type, tex, rec);
     }
